fix: resolve 34 pre-existing test failures
- Use config('app.fallback_locale') instead of hardcoded 'id' in
  AppServiceProvider to respect locale configuration in tests
- Fix Currency::getDefault() return type from Closure|self to ?self
- Fix CouponService::calculateDiscount() float return type for fixed coupons
- Add try-catch for PermissionDoesNotExist in SettingPage and
  SendNewsletter canAccess() when permissions table is empty
- Fix LocalizationTest homepage_blocks setting key mismatch
- Seed Currency in CouponServiceTest to prevent UNIQUE constraint violations

+++ app/Services/CouponService.php
@@ -78,7 +78,7 @@ final class CouponService
     public function calculateDiscount(Coupon $coupon, int $subtotal): int
     {
         return match ($coupon->type->value) {
-            'fixed' => min($coupon->value, $subtotal),
+            'fixed' => (int) min($coupon->value, $subtotal),
             'percentage' => (int) round($subtotal * $coupon->value / 100),
             default => 0,
         };
